requirements.txt
hikari~=2.0.0.dev122
attrs~=23.1.0
